HTTP headers, SwiftLint changes
